+++ nextgisweb/core/__init__.py
@@ -377,6 +377,8 @@ class CoreComponent(
         Option('locale.default', default='en'),
         Option('locale.available', list, default=None),
         Option('locale.external_path', default=None),
+        Option('locale.poeditor.project_id', str, default=None),
+        Option('locale.poeditor.api_token', str, default=None),
 
         # Other deployment settings
         Option('support_url', default="https://nextgis.com/contact/"),