+++ nextgisweb/i18n/script.py
@@ -6,12 +6,16 @@ import os
 import os.path
 import logging
 import json
+import re
 import six
 from importlib import import_module
+from packaging import version as pkg_version
 from argparse import ArgumentParser, Namespace
 from pkg_resources import resource_filename
-from collections import OrderedDict
+from collections import OrderedDict, defaultdict
 from functools import partial
+from tempfile import NamedTemporaryFile
+from time import sleep
 
 from babel.messages.catalog import Catalog
 from babel.messages.frontend import parse_mapping
@@ -19,6 +23,7 @@ from babel.messages.extract import extract_from_dir
 from babel.messages.pofile import write_po, read_po
 from babel.messages.mofile import write_mo
 from attr import attrs, attrib, asdict
+from poeditor import POEditorAPI
 
 from ..compat import Path
 from ..package import pkginfo
@@ -126,7 +131,7 @@ def components_and_locales(args, work_in_progress=False):
                     candidate = fn.with_suffix('').name
                     if candidate not in locales:
                         locales.append(candidate)
-        
+      
         locales.sort()
         logger.debug("Locale list for [%s] = %s", comp_id, ' '.join(locales))
         yield comp_id, locales
@@ -391,6 +396,194 @@ def cmd_stat(args):
         print_records(data, ('package', ), title='PACKAGE SUMMARY')
 
 
+def cmd_poeditor_sync(args):
+    opts = env.core.options.with_prefix('locale.poeditor')
+
+    if 'project_id' not in opts:
+        raise RuntimeError("POEditor project ID isn't set!")
+    poeditor_project_id = opts['project_id']
+
+    if 'api_token' not in opts:
+        raise RuntimeError("POEditor API token isn't set!")
+    poeditor_api_token = opts['api_token']
+
+    client = POEditorAPI(api_token=poeditor_api_token)
+
+    # Package versions are stored as part of the project's description in the POEditor
+    description =  POEditorDescription(client, poeditor_project_id)
+    local_ver = pkginfo.packages[args.package].version
+    if args.package in description.packages:
+        remote_ver = description.packages[args.package]
+
+        if pkg_version.parse(local_ver) < pkg_version.parse(remote_ver):
+            raise RuntimeError(
+                "Version of the '%s' package must be not lower than %s in order to use translations from the POEditor "
+                "(current version: %s)."
+                % (args.package, remote_ver, local_ver)
+            )
+
+    # Update local po-files
+    poeditor_terms = {}
+    reference_catalogs = {}
+    for comp_id, locales in components_and_locales(args, work_in_progress=True):
+        cmd_update(Namespace(
+            package=pkginfo.comp_pkg(comp_id),
+            component=[comp_id, ],
+            locale=[l for l in locales if l != 'ru'],
+            extract=args.extract,
+            force=False))
+
+        for locale in locales:
+            if locale == 'ru':
+                po_path = catalog_filename(comp_id, locale)
+                if po_path.exists():
+                    ref_catalog = reference_catalogs.get(locale)
+                    if ref_catalog is None:
+                        ref_catalog = Catalog(locale=locale)
+                        reference_catalogs[locale] = ref_catalog
+                    with po_path.open('r') as po_fd:
+                        for m in read_po(po_fd, locale=locale):
+                            m.context = comp_id
+                            ref_catalog[m.id] = m
+                continue
+
+            terms = poeditor_terms.get(locale)
+            if not terms:
+                logger.debug("Fetching translations from POEditor for locale [%s]...", locale)
+                language_code = locale.replace('_', '-').lower()
+                terms = client.view_project_terms(poeditor_project_id, language_code=language_code)
+                poeditor_terms[locale] = terms
+
+            # Filter terms by context
+            terms = [term for term in terms if comp_id == term['context']]
+
+            po_path = catalog_filename(comp_id, locale)
+
+            if not po_path.exists():
+                continue
+
+            with po_path.open('r') as po_fd:
+                po = read_po(po_fd, locale=locale)
+
+            updated = 0
+
+            for msg in po:
+                for term in terms:
+                    if msg.id == term['term']:
+                        cur_tr = msg.string
+                        new_tr = term['translation']['content']
+
+                        if cur_tr != new_tr:
+                            msg.string = new_tr
+                            updated += 1
+
+                        break
+
+            if updated != 0:
+                with io.open(str(po_path), 'wb') as fd:
+                    write_po(fd, po, width=80, omit_header=True)
+
+                logger.info(
+                    "%d messages translated for component [%s] locale [%s]",
+                    updated, comp_id, locale)
+
+    # Synchronize terms
+    remote_terms = defaultdict(set)
+    target_state = defaultdict(dict)
+
+    for item in client.view_project_terms(poeditor_project_id):
+        remote_terms[item['term']].add(item['context'])
+
+    comps = []
+    for comp_id, _ in components_and_locales(args, work_in_progress=True):
+        pot_path = catalog_filename(comp_id, '', ext='pot')
+        with pot_path.open('r') as fd:
+            pot = read_po(fd)
+        for msg in pot:
+            if msg.id == '':
+                continue
+            target_state[msg.id][comp_id] = True
+        comps.append(comp_id)
+
+    for comp_id in comps:
+        for msg_id in target_state:
+            if comp_id not in target_state[msg_id]:
+                target_state[msg_id][comp_id] = False
+
+    terms_to_add = []
+    terms_to_del = []
+    for msg_id in target_state:
+        if not msg_id in remote_terms:
+            for comp_id in target_state[msg_id]:
+                if target_state[msg_id][comp_id]:
+                    terms_to_add.append(dict(term=msg_id, context=comp_id))
+            continue
+
+        for comp_id in target_state[msg_id]:
+            if target_state[msg_id][comp_id] and comp_id not in remote_terms[msg_id]:
+                terms_to_add.append(dict(term=msg_id, context=comp_id))
+            if not target_state[msg_id][comp_id] and comp_id in remote_terms[msg_id]:
+                terms_to_del.append(dict(term=msg_id, context=comp_id))
+
+    if len(terms_to_add) > 0 and args.no_dry_run:
+        client.add_terms(poeditor_project_id, terms_to_add)
+    logger.info("%d messages added to the POEditor", len(terms_to_add))
+
+    if len(terms_to_del) > 0 and args.no_dry_run:
+        client.delete_terms(poeditor_project_id, terms_to_del)
+    logger.info("%d messages deleted from the POEditor", len(terms_to_del))
+
+    if len(reference_catalogs) > 0:
+        logger.info(
+            "Upload the following reference translations to POEditor: " +
+            ", ".join(reference_catalogs.keys()))
+        if args.no_dry_run:
+            wait_for_rate_limit = False
+            for locale, catalog in reference_catalogs.items():
+                with NamedTemporaryFile(suffix='.po') as fd:
+                    write_po(fd, catalog, width=80, omit_header=True)
+                    fd.flush()
+                    logger.debug(
+                        "Uploading %s reference translation...",
+                        locale)
+
+                    # Free account allows doing 1 upload per 20 seconds
+                    sleep(20 if wait_for_rate_limit else 0)
+                    wait_for_rate_limit = True
+
+                    client.update_terms_translations(
+                        project_id=poeditor_project_id,
+                        language_code=locale.replace('-', '_'),
+                        overwrite=True, sync_terms=False,
+                        file_path=fd.name)
+
+    if len(terms_to_add) > 0 or len(terms_to_del) > 0:
+        if args.no_dry_run:
+            description.update(args.package, local_ver)
+        logger.info(
+            "Set '%s' package version to %s in the POEditor.", args.package, local_ver
+        )
+
+    # TODO: Russian language as a reference
+
+
+class POEditorDescription:
+    def __init__(self, client, project_id):
+        self.client = client
+        self.project_id = project_id
+        self.packages = []
+        self.populate()
+
+    def populate(self):
+        details =  self.client.view_project_details(self.project_id)
+        self.packages = json.loads(re.search('\((.+?)\)', details['description']).group(1))
+
+    def update(self, package, version):
+        self.packages[package] = version
+        description = '[//]: # (%s)' % (json.dumps(self.packages))
+        self.client.update_project(project_id=self.project_id, description=description)
+
+
 def main(argv=sys.argv):
     logging.basicConfig(level=logging.INFO)
 
@@ -423,6 +616,13 @@ def main(argv=sys.argv):
     pstat.add_argument('--work-in-progress', action='store_true', default=False)
     pstat.add_argument('--json', action='store_true', default=False)
     pstat.set_defaults(func=cmd_stat)
+
+    ppoeditor_pull = subparsers.add_parser('poeditor-sync')
+    ppoeditor_pull.add_argument('component', nargs='*')
+    ppoeditor_pull.add_argument('--extract', action='store_true', default=False)
+    ppoeditor_pull.add_argument('--locale', default=[], action='append')
+    ppoeditor_pull.add_argument('--no-dry-run', action='store_true', default=False)
+    ppoeditor_pull.set_defaults(func=cmd_poeditor_sync)
 
     args = parser.parse_args(argv[1:])